+++ presentation/src/main/java/com/pat/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.pat.presentation.ui.ui.theme
+package com.pat.presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
